refactor: 승자찾기 로직과 UI 로직 분리
